Add files via upload
In the final defense we build this project which is based on oxygen cylinder management system. Where you can get medical tools specially oxygen cylinder. Actually me and my team brought the idea after covid19 where lots of patients died lack of medical equipment like oxygen cylinder. This is one of the best project we demonstrate for the final defense.